retry update

+++ smqtt-common/src/main/java/io/github/quickmsg/common/message/mqtt/PublishAckMessage.java
@@ -15,7 +15,7 @@ import lombok.Data;
 @Data
 public class PublishAckMessage implements Message {
 
-    private long messageId;
+    private int messageId;
 
     private long timestamp;
